feat(community): 建立社区应用服务与错误码
PostService / CommentService / ReactionService / TagService，以及
PostCardView / PostDetailView / CommentView / PostCommand / FeedQuery /
ReactionResult 六个应用层模型。

归属校验：编辑与删除都在服务层重新查帖子并比对 author_id，失败一律 404 而非 403。
403 等于告诉调用方"这条帖子存在，只是不归你"，那就是一个内容存在性探测器。
与 Phase 02 的"下线他人会话返回 404"是同一条规则。

详情读取刻意不是只读事务：它会顺带记录一次浏览（post_view_daily 去重 + 计数）。
放在同一事务里，是为了避免为了记一次浏览就挂起当前事务、再从连接池取第二条连接 ——
详情是最高频的读路径，连接切换的成本会直接落在它身上。
代价（记账失败会连带影响读取）用 try/catch 兜住：计数是次要信息，
不该让用户看不到内容。返回的浏览数会 +1 修正，否则用户会看到"我刚看完，数字没变"。

评论的层级约束只能由服务层保证：外键只能保证 parent_id 指向一条存在的评论，
无法保证那条评论本身是顶层。漏掉检查就会悄悄产生第三层，而回复查询只查一层，
那些内容会"存在于库里但界面上永远看不到"。检查点三条缺一不可：
父评论存在、属于同一帖子、自身是顶层 —— 少了"同一帖子"，
回复就会出现在 A 帖的评论区而它回复的内容在 B 帖。

互动幂等：两个方向都直接执行、由唯一键裁决，撞键即代表"已是该状态"且不报错。
"先查再决定"在并发下两个请求都会查到"没有"，然后其中一个必然失败，
并把它变成一次用户可见的错误。
计数调整写成按类型分支的两个方法，而不是拼列名 —— 列名无法用占位符传递，
拼接就意味着那一处是注入面。

标签"查找或创建"容忍撞唯一键：两个用户同时用同一个新标签发布，
一个成功一个撞 uk_tag_slug，这是并发下的正常路径而不是缺陷。
不加重试、不加锁，交回数据库裁决。去重发生在规范化之后：
用户同时填了 Java 与 java 时，按原字符串去重会让两者解析到同一 slug，
最终在关联表主键 (post_id, tag_id) 上撞键报错。

新增错误码 40030/40031/40032（community 段），并明确 40031 与 40000 的分工：
前者管"业务策略"（上限来自配置、会被压测调整），后者管"字段格式"（Bean Validation）。
合成一个码之后，前端就无法区分"少填了字段"与"写得太长"。

新增审计动作 CONTENT_POST_CREATE / UPDATE / DELETE / COMMENT_DELETE。
内容的创建也记审计，而不只记删除：审计在这里要回答"这条内容是什么时候、
由谁、通过哪次请求产生的"，而仅凭 created_at 与 author_id 无法区分
"用户自己发的"与"某个接口被滥用代发的"。

验证：./mvnw -B -o compile 通过。

+++ src/main/java/ai/camphub/community/infrastructure/PostTagMapper.java
@@ -19,11 +19,15 @@ public interface PostTagMapper {
      *
      * <p>调用方在写入前会先清空该帖的旧关联，因此这里不需要处理重复键。
      *
+     * <p>标签 id 在 Java 侧统一用 {@code long} 承载（{@code INT UNSIGNED} 的上限超出
+     * {@code int} 的正数范围，而项目其它模块的主键也都是 long）。让标识类型在全项目一致，
+     * 比在这里省几个字节重要 —— 混用会在跨模块边界反复产生无意义的窄化转换。
+     *
      * @param postId 帖子自增主键
      * @param tagIds 标签自增主键集合（非空）
      * @return 影响行数
      */
-    int insertBatch(@Param("postId") long postId, @Param("tagIds") Collection<Integer> tagIds);
+    int insertBatch(@Param("postId") long postId, @Param("tagIds") Collection<Long> tagIds);
 
     /**
      * 清除某帖的全部标签关联。
